chore(example): rename status attestation screen
